Fixes to api_v2 plus fully containerized solution for quads

+++ quads/tools/create_input.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
 
 import os
+import pathlib
 import re
-from helpers import quads_load_config
-from tools.foreman import Foreman
-from quads import Quads
+import requests
+from quads.helpers import quads_load_config
+from quads.tools.foreman import Foreman
 
 conf_file = os.path.join(os.path.dirname(__file__), "../../conf/quads.yml")
 conf = quads_load_config(conf_file)
@@ -23,19 +24,26 @@ HEADERS = [
     "Graph"
 ]
 
+API = 'v2'
+
 
 def consolidate_ipmi_data(_host, _path, _value):
-    _file_path = os.path.join(conf["data_dir"], "ipmi", _host, _path)
+    ipmi_path = os.path.join(conf["data_dir"], "ipmi")
+    host_path = os.path.join(ipmi_path, _host)
+    _file_path = os.path.join(host_path, _path)
     try:
         with open(_file_path, "r+") as _ipmi_file:
             mac = _ipmi_file.read()
-            if not mac:
+            if not mac and _value:
                 _ipmi_file.seek(0)
                 _ipmi_file.write(_value)
                 _ipmi_file.truncate()
-    except IOError:
+    except (IOError, FileNotFoundError):
+        if not os.path.exists(host_path):
+            pathlib.Path(host_path).mkdir(parents=True, exist_ok=True)
         with open(_file_path, "w") as _ipmi_file:
-            _ipmi_file.write(_value)
+            value = _value if _value else ""
+            _ipmi_file.write(value)
 
 
 def render_header(_rack):
@@ -45,16 +53,24 @@ def render_header(_rack):
     return "\n".join([h, "", h1, h2])
 
 
-def render_row(_quads, _host, _properties):
+def render_row(_host, _properties):
+    api_url = os.path.join(conf['quads_base_url'], 'api', API)
     u_loc = _host.split("-")[2][1:]
     node = _host[5:]
     owner = ""
     cloud = ""
-    workload = _quads.query_host_cloud(node, None)
-    if workload:
-        cloud_owner = _quads.get_owners(workload)
-        owner = cloud_owner[0][workload]
-        cloud = "[%s](/assignments/#%s)" % (workload, workload)
+    host_url = os.path.join(api_url, "host?name=%s" % node)
+    _response = requests.get(host_url)
+    if _response.status_code == 200:
+        host_data = _response.json()
+        if "cloud" in host_data:
+            cloud_url = os.path.join(api_url, "cloud?name=%s" % host_data["cloud"])
+            _cloud_response = requests.get(cloud_url)
+            if _cloud_response.status_code == 200:
+                cloud_data = _cloud_response.json()
+                if "owner" in cloud_data:
+                    owner = cloud_data["owner"]
+            cloud = "[%s](/assignments/#%s)" % (host_data["cloud"], host_data["cloud"])
 
     # TODO: figure out what to put on grafana field
     grafana = ""
@@ -82,16 +98,6 @@ def rack_has_hosts(rack, hosts):
 
 
 def main():
-    default_config = conf["data_dir"] + "/schedule.yaml"
-    default_state_dir = conf["data_dir"] + "/state"
-    default_move_command = "/bin/echo"
-
-    quads = Quads(
-        default_config,
-        default_state_dir,
-        default_move_command,
-        None, False, False, False,
-    )
     foreman = Foreman(
         conf["foreman_api_url"],
         conf["foreman_username"],
@@ -103,8 +109,8 @@ def main():
     hosts = {}
     for host, properties in all_hosts.items():
         if "mgmt" in host and not blacklist.search(host):
-            properties["host_ip"] = all_hosts.get(host[5:], {"ip": None})["ip"]
-            properties["host_mac"] = all_hosts.get(host[5:], {"mac": None})["mac"]
+            properties["host_ip"] = all_hosts.get(host[5:], {"ip": ""})["ip"]
+            properties["host_mac"] = all_hosts.get(host[5:], {"mac": ""})["mac"]
             consolidate_ipmi_data(host[5:], "macaddr", properties["host_mac"])
             consolidate_ipmi_data(host[5:], "oobmacaddr", properties["mac"])
             svctag_file = os.path.join(conf["data_dir"], "ipmi", host[5:], "svctag")
@@ -117,6 +123,9 @@ def main():
 
     _full_path = os.path.join(conf["wp_wiki_git_repo_path"], "main.md")
 
+    if not os.path.exists(conf["wp_wiki_git_repo_path"]):
+        pathlib.Path(conf["wp_wiki_git_repo_path"]).mkdir(parents=True, exist_ok=True)
+
     with open(_full_path, "w") as _f:
         _f.seek(0)
         for rack in conf["racks"].split():
@@ -125,7 +134,7 @@ def main():
 
                 for host, properties in hosts.items():
                     if rack in host:
-                        _f.write(render_row(quads, host, properties))
+                        _f.write(render_row(host, properties))
                 _f.write("\n")
 
         _f.truncate()
